+++ secrets/hash_test.go
@@ -1,0 +1,8 @@
+package secrets
+
+import "testing"
+
+func TestHash(t *testing.T) {
+	// TODO: HashPassword
+	// TODO: MatchPassword
+}